* fixed compile problems if SPLIT_PONG was undefined * more cleanup * moved TIMED_K's to broken section of conf

+++ tools/fixklines.c
@@ -240,7 +240,7 @@ int host_is_legal_dline(char *host_name)
           else
             return(NO);
         }
-      else if(!isdigit(*host_name))
+      else if(!isdigit((int) *host_name))
         return(NO);
       host_name++;
     }
